refactor: move Style3DSolver kernels to `kernels.py`
